pyproject.toml
[project]
name = "hr-agent"
version = "0.1.0"
description = "VanaciRetain - Agentic RAG System for HR Teams"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "VanaciPrime" }]
dependencies = [
    # Agent Framework
    "langgraph>=0.4.0,<2.0.0",
    "langchain>=0.3.0,<2.0.0",
    "langchain-core>=0.3.0,<2.0.0",
    "langchain-community>=0.3.0,<2.0.0",
    "langchain-groq>=0.2.0,<2.0.0",
    "langchain-text-splitters>=0.3.0,<2.0.0",
    # Document Loading
    "pymupdf>=1.24.0,<2.0.0",
    # Embeddings (pinned for Intel Mac compatibility)
    "torch>=2.6.0",
    "sentence-transformers>=3.0.0,<4.0.0",
    "transformers>=4.44.0,<4.50.0",
    "langchain-huggingface>=0.1.0,<2.0.0",
    # Vector Database
    "qdrant-client>=1.12.0,<2.0.0",
    "langchain-qdrant>=0.2.0,<2.0.0",
    # Structured Database
    "supabase>=2.0.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # RAG Evaluation
    "ragas>=0.2.0,<0.4.0",
    # ML Models
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0,<2.0.0",
    # MLOps
    "mlflow>=2.18.0",
    # Re-ranking
    "flashrank>=0.2.0",
    # Rate limit retry
    "groq>=0.37.0",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "agents", "rag", "models", "mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.uv]
constraint-dependencies = [
    "onnxruntime>=1.21.0",
    "pyarrow>=18.0.0",
]
